Pré-Finalização

+++ src/Membros.js
@@ -1,12 +1,25 @@
+// Trabalho 1 - Desenvolvimento de Sistemas Web - T01 - 2022.1
+    
+// Alunos:
+// Denilson Luiz Amaro Martins - 2020031625
+// Pedro Henrique Azevedo do Prado - 2022001536
+
+// Professor:
+// Eduardo Ribeiro Felipe
+
 import React, { useEffect, useState } from "react";
 import axios from "axios";
 import imgEdit from "./images/imgEdit.ico";
 import imgDelete from "./images/imgDelete.ico";
 import decolar from "./images/decolar.ico";
-import pousar from "./images/pousar.ico"
+import pousar from "./images/pousar.ico";
+import decolarInverso from "./images/decolarInverso.jpeg";
+import pousarInverso from "./images/pousarInverso.jpeg";
 import "./Membros.css";
 
 export default function Membros() {
+  {/* Definindo variáveis e métodos para estas */}
+
   const [membros, setMembros] = useState([]);
   const [id, setId] = useState("");
   const [nome, setNome] = useState("");
@@ -19,14 +32,14 @@ export default function Membros() {
   const [cargo, setCargo] = useState("");
   const [tipo, setTipo] = useState("");
 
-  // const url = "https://backend-n3w118fyk-denilsonluizm.vercel.app/";
+  {/* Inserindo a URL do banco de dados */}
 
   const url = "https://backend-gnx9kw2l3-denilsonluizm.vercel.app/";
 
-  // const url = "http://localhost:8081/";
+  {/* Puxando dados atuais do banco de dados */}
 
   useEffect(() => {
-    fetch(url + "membros", { headers: { 'Access-Control-Allow-Origin': '*' } })
+    fetch(url + "membros")
       .then((response) => response.json())
       .then((data) => setMembros(data))
       .catch((err) => console.log(err));
@@ -35,6 +48,8 @@ export default function Membros() {
   function novosDados() {
     setTipo("novo");
   }
+
+  {/* Limpando as caixas de preenchimento de informações de membros */}
 
   function limparDados() {
     setId("");
@@ -48,6 +63,8 @@ export default function Membros() {
     setCargo("");
     setTipo("");
   }
+
+  {/* Colocando informações nas caixas de preenchimento de informações de membros */}
 
   function editarDados(cod) {
     let membro = membros.find((item) => item.id === cod);
@@ -64,12 +81,22 @@ export default function Membros() {
     setCargo(cargo);
   }
 
+  {/* Função ativada quando é solicitada a EXCLUSÃO de um membro - com confirmações e avisos. */}
+
   function apagarDados(cod) {
-    axios.delete(url + "membros/" + cod).then(() => {
-      //atualizar a lista
-      setMembros(membros.filter((item) => item.id !== cod));
-    });
+    if(window.confirm('Você tem certeza que deseja apagar o membro?')){
+      axios.delete(url + "membros/" + cod).then(() => {
+        //atualizar a lista
+        setMembros(membros.filter((item) => item.id !== cod));
+      });
+      alert("Você APAGOU este membro!");
+    }else{
+      alert("Você NÃO apagou este membro!");
+    }
+    
   }
+
+  {/* Adicionando informações de um novo membro cadastrado */}
 
   function atualizaListaComNovoMembro(response) {
     let { id, nome, matricula, curso, setor, email, telefone, status, cargo } = response.data;
@@ -79,6 +106,8 @@ export default function Membros() {
     setMembros(members);
     limparDados("");
   }
+
+  {/* Atualizando informações de um membro já cadastrado */}
 
   function atualizaListaMembroEditado(response) {
     let { id } = response.data;
@@ -95,6 +124,8 @@ export default function Membros() {
     setMembros(members);
     limparDados("");
   }
+
+  {/* Função que faz a requisição POST para cadastrar as informações de um novo membro no banco de dados */}
 
   function gravaDados() {
     if (nome !== "" && matricula !== "" && email !== "") {
@@ -113,6 +144,7 @@ export default function Membros() {
           .then((response) => atualizaListaComNovoMembro(response))
           .catch((err) => console.log(err));
       } else if (tipo === "editar") {
+        {/* Função que faz a requisição PUT para cadastrar as novas informações de um membro já existente no banco de dados */}
         axios
           .put(url + "membros/" + id, {
             id: id,
@@ -133,6 +165,8 @@ export default function Membros() {
     }
   }
 
+
+  {/* Interface intuitiva de cadastro, atualização e remoção de membros */}
   return (
     <div className="App">
       <br></br>
@@ -232,14 +266,6 @@ export default function Membros() {
               setTelefone(e.target.value);
             }}
           />
-          {/* <input
-            type="text"
-            name="txtStatus"
-            value={status}
-            onChange={(e) => {
-              setStatus(e.target.value);
-            }}
-          /> */}
           <br></br>
           <br></br>
           Status:
@@ -281,29 +307,58 @@ export default function Membros() {
       ) : (
         false
       )}
+      {/* Função flecha que mostra, de forma cadenciada, os membros já cadastrados no banco de dados */}
       {membros
         ? membros.map((item) => {
           return (
             <div key={item.id}>
               <div className="linha">
-                {item.id} - {item.nome} - {item.matricula} - {item.curso} - {item.setor} - {item.email} - {item.telefone} - {item.status} - {item.cargo}{" "}
+                <b>ID:</b> {item.id} <br></br>
+                <b>Nome:</b> {item.nome} <br></br>
+                <b>Matrícula:</b> {item.matricula} <br></br>
+                <b>Curso:</b> {item.curso} <br></br>
+                <b>Setor:</b> {item.setor} <br></br>
+                <b>E-mail:</b> {item.email} <br></br>
+                <b>Telefone:</b> {item.telefone} <br></br>
+                <b>Status:</b> {item.status} <br></br>
+                <b>Cargo:</b> {item.cargo}{" "} <br></br>
+                {/* Botão que dá a possibilidade de alterar informações de membros do banco de dados */}
+                <img
+                  alt="Editar"
+                  src={decolarInverso}
+                  id={item.id}
+                  height={20}
+                  width={20}
+                />
+                <button onClick={(e) => editarDados(item.id)}>Editar</button>
                 <img
                   alt="Editar"
                   src={decolar}
                   id={item.id}
                   height={20}
                   width={20}
-                  onClick={(e) => editarDados(item.id)}
                 />
                 {"  "}
+                <br></br>
+                {/* Botão que dá a possibilidade de apagar membros do banco de dados*/}
+                <img
+                  alt="Apagar"
+                  src={pousarInverso}
+                  id={item.id}
+                  height={20}
+                  width={20}
+                />
+                <button alert onClick={(e) => apagarDados(item.id)}>Apagar</button>
                 <img
                   alt="Apagar"
                   src={pousar}
                   id={item.id}
                   height={20}
                   width={20}
-                  onClick={(e) => apagarDados(item.id)}
                 />
+                <br></br>
+                <br></br>
+                <br></br>
               </div>
             </div>
           );
